Update patch set 2
Patch Set 2:

Well, it really needs the whole of https://review.lineageos.org/#/q/topic:board-flags-ftw to work... Whatever that initial pointer inserts, it doesn't look like a flat Go struct, so they had to nest, which cascades into ugly changes for us. Maybe we can find another way to keep it flat with the JSON marshalling stuff, but as of now we have only one new Go var, so not much incentive to make it live in vendor.

Patch-set: 2